pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-retail-loss-prevention-intelligence-platform"
version = "0.1.0"
description = "Agentic retail shrinkage detection demo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "httpx>=0.28.0",
  "pydantic>=2.8.0",
  "numpy>=1.26.0",
  "pillow>=10.4.0",
  "ultralytics>=8.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
